Fix a bug with complicated version specifications
Patched versions and unaffected versions can sometimes have compound
version specifications. That is, versions that are specified like so:
`~> 4.2.5, > 4.2.5.1`. In order for these compound version
specifications to be handled correctly with bundler-advise, we need to
split them when calling `Gem:Requirement.create`, which is what this
commit does.

It also updates the test suite to handle these compound version
specifications.