GemAdviser sets gem_spec on Advisory instances.
This is a bit hack-ish, 'breaking' the interface of the Advisory class
by tacking on a field that's not part of the Advisory at all, but links
in the related gem spec read from the lockfile.

But it helps follow semver and is a small enough violation, I'm cool
with it.