Remove lockfile.
Many folks don't like this normally, I'm usually ambivalent or prefer
to keep things locked down across checkouts ... but now with trying to
support Bundler 1 and 2, Bundler "helps" by identifying a mismatch in
the Bundled version in the lockfile and what's installed locally, which
is a fine use-case for end-users, but gets in the way for what I'm
trying to do here with the build matrix. So, let's see how this goes.